+++ test/integration/cookbooks/testrig/metadata.rb
@@ -6,7 +6,7 @@ description      'Test cookbook for chef-kubernetes'
 long_description 'Test cookbook for chef-kubernetes'
 version          '0.0.1'
 
-supports 'ubuntu', '>= 14.04'
+supports 'ubuntu', '>= 16.04'
 
 depends 'kubernetes'